Fix to work with MPS 2018.2, bump version from 0.1 to 0.2; now incompatible with MPS 2018.1
Reason for incompatibility: now using DefaultModelRoot#createDescriptor()
which was introduced in 2018.2. Ctor call new DefaultModelRoot() still can
be used, but will throw an exception, because solution descriptor is created
too early, before mps platform is initialised (happens in GeneratorWorker).
In 2018.2 that ctor tries to access DataSourceFactoryRuleService#getInstance()
which won't be initialised yet.